requirements.txt
yt-dlp>=2023.11.16
Pillow>=8.0.0
requests>=2.25.0
ffmpeg-python>=0.2.0 